+++ vertigo-elasticsearch-connector/src/main/java/io/vertigo/connectors/elasticsearch/ElasticSearchFeatures.java
@@ -23,6 +23,7 @@ import io.vertigo.core.param.Param;
 
 /**
  * Defines ElasticSearchConnector module.
+ *
  * @author npiedeloup
  */
 public final class ElasticSearchFeatures extends Features<ElasticSearchFeatures> {
@@ -34,31 +35,22 @@ public final class ElasticSearchFeatures extends Features<ElasticSearchFeatures>
 		super("vertigo-elasticsearch-connector");
 	}
 
+	@Feature("rest")
+	public ElasticSearchFeatures withRest(final Param... params) {
+		getModuleConfigBuilder()
+				.addConnector(RestElasticSearchConnector.class, params);
+		return this;
+	}
+
 	@Feature("embeddedServer")
 	public ElasticSearchFeatures withEmbeddedServer(final Param... params) {
+		/**
+		 * To use EmbeddedServer with org.testcontainers, we need to declare a varEnv DOCKER_HOST to point on docker socket (like DOCKER_HOST=tcp://localhost:2375)
+		 * To active socket on a docket with WSL, you shoud edit "/lib/systemd/system/docker.service" and add this line in [Service] section:
+		 * ExecStart=/usr/bin/dockerd -H fd:// -H tcp://127.0.0.1:2375 --containerd=/run/containerd/containerd.sock
+		 */
 		getModuleConfigBuilder()
 				.addComponent(EmbeddedElasticSearchServer.class, params);
-		return this;
-	}
-
-	@Feature("transport")
-	public ElasticSearchFeatures withTransport(final Param... params) {
-		getModuleConfigBuilder()
-				.addConnector(TransportElasticSearchConnector.class, params);
-		return this;
-	}
-
-	@Feature("node")
-	public ElasticSearchFeatures withNode(final Param... params) {
-		getModuleConfigBuilder()
-				.addConnector(NodeElasticSearchConnector.class, params);
-		return this;
-	}
-
-	@Feature("restHL")
-	public ElasticSearchFeatures withRestHL(final Param... params) {
-		getModuleConfigBuilder()
-				.addConnector(RestHighLevelElasticSearchConnector.class, params);
 		return this;
 	}