+++ vertigo-oidc-connector/src/main/java/io/vertigo/connectors/oidc/IOIDCDeploymentConnector.java
@@ -1,0 +1,7 @@
+package io.vertigo.connectors.oidc;
+
+import io.vertigo.core.node.component.Connector;
+
+public interface IOIDCDeploymentConnector extends Connector<IOIDCClient> {
+
+}