+++ vertigo-oidc-connector/src/main/java/io/vertigo/connectors/oidc/OIDCDeploymentConnector.java
@@ -22,7 +22,6 @@ import java.util.Optional;
 import javax.inject.Inject;
 
 import io.vertigo.core.lang.Assertion;
-import io.vertigo.core.node.component.Connector;
 import io.vertigo.core.param.ParamValue;
 import io.vertigo.core.resource.ResourceManager;
 
@@ -31,7 +30,7 @@ import io.vertigo.core.resource.ResourceManager;
  *
  * @author skerdudou
  */
-public class OIDCDeploymentConnector implements Connector<OIDCClient> {
+public class OIDCDeploymentConnector implements IOIDCDeploymentConnector {
 
 	private final OIDCClient oidcClient;
 	private final String connectorName;
@@ -55,6 +54,8 @@ public class OIDCDeploymentConnector implements Connector<OIDCClient> {
 			@ParamValue("logoutRedirectUriParamName") final Optional<String> logoutRedirectUriParamNameOpt,
 			@ParamValue("logoutIdParamName") final Optional<String> logoutIdParamNameOpt,
 			@ParamValue("localeParamName") final Optional<String> localeParamNameOpt,
+			@ParamValue("proxyHost") final Optional<String> proxyHostOpt,
+			@ParamValue("proxyPort") final Optional<Integer> proxyPortOpt,
 			final ResourceManager resourceManager) {
 
 		Assertion.check()
@@ -74,11 +75,13 @@ public class OIDCDeploymentConnector implements Connector<OIDCClient> {
 				.isNotNull(trustStorePasswordOpt)
 				.isNotNull(logoutRedirectUriParamNameOpt)
 				.isNotNull(logoutIdParamNameOpt)
-				.isNotNull(localeParamNameOpt);
+				.isNotNull(localeParamNameOpt)
+				.isNotNull(proxyHostOpt)
+				.isNotNull(proxyPortOpt);
 		//---
 		connectorName = connectorNameOpt.orElse("main");
 
-		final OIDCParameters oidcParameters = new OIDCParameters(
+		final var oidcParameters = new OIDCParameters(
 				clientName,
 				clientSecretOpt,
 				oidcUrl,
@@ -94,7 +97,9 @@ public class OIDCDeploymentConnector implements Connector<OIDCClient> {
 				localeParamNameOpt,
 				dontFailAtStartupOpt.orElse(Boolean.FALSE),
 				trustStoreUrlOpt,
-				trustStorePasswordOpt);
+				trustStorePasswordOpt,
+				proxyHostOpt,
+				proxyPortOpt.orElse(3128));// default port for http proxy
 
 		oidcClient = new OIDCClient(oidcParameters);
 	}